Move symbols from types to utils
