+++ svglab/parse.py
@@ -1,12 +1,16 @@
 import collections
 import itertools
-from typing import Final, cast
+from typing import Final, Literal, TypeAlias, cast
 
 import bs4
 
-from svglab import elements, types, utils
+from svglab import elements, utils
 
-DEFAULT_PARSER: Final[types.Parser] = "lxml-xml"
+Parser: TypeAlias = Literal["html.parser", "lxml", "lxml-xml", "html5lib"]
+""" Type for parsers supported by BeautifulSoup. """
+
+
+DEFAULT_PARSER: Final[Parser] = "lxml-xml"
 """The default parser to use when parsing SVG documents."""
 
 
@@ -135,11 +139,11 @@ def convert_element(
 def parse_svg(
     markup: str
     | bytes
-    | types.SupportsRead[str]
-    | types.SupportsRead[bytes],
+    | utils.SupportsRead[str]
+    | utils.SupportsRead[bytes],
     /,
     *,
-    parser: types.Parser = DEFAULT_PARSER,
+    parser: Parser = DEFAULT_PARSER,
 ) -> elements.Svg:
     """Parse an SVG document.
